+++ src/main/java/com/powsybl/openloadflow/network/LfElement.java
@@ -19,6 +19,10 @@ public interface LfElement {
 
     void setNum(int num);
 
+    boolean isDisabled();
+
+    void setDisabled(boolean disabled);
+
     LfNetwork getNetwork();
 
     Object getUserObject();
